refactor reorderGroupPopulations to util class
